fix typeConverter error for Wind.kt

+++ app/src/main/java/com/jobinlawrance/weather/data/network/model/Wind.kt
@@ -1,6 +1,6 @@
 package com.jobinlawrance.weather.data.network.model
 
 data class Wind(
-	val deg: Int? = null,
+	val deg: Double? = null,
 	val speed: Double? = null
 )
